feat: allowing conversion of ChainedDictionary to Dictionary
Added method MakeDictionary, which will create a Dictionary instance based
on the ChainedDictionary. The instance will not be a undetermined chain
any longer, but will be an instance of the chain of Dictionary informed
in the generic parameters. Due to C# limitations, we can't validate that
instance in compile time, so, some safe guards are set during creation
to make sure the chain is compatible with the ChainedDictionary instance, ie,
all the keys in the path have compatible type